+++ interpreter_webassembly/src/flag_register.hpp
@@ -4,17 +4,17 @@ private :
   bool overflow_flag;
   bool zero_flag;
   bool sign_flag;
-  void set_sf_zf(int16_t value);
+  void set_sf_zf(int value);
 
 public :
   FlagRegister();
   bool of();
   bool zf();
   bool sf();
-  void set(int16_t value);
-  void set_with_overflow_flag(int16_t value, bool overflow_flag);
-  void set_logical(uint16_t value);
-  void set_by_cpa(int16_t value);
-  void set_logical_by_cpl(int16_t value);
+  void set(int value);
+  void set_with_overflow_flag(int value, bool overflow_flag);
+  void set_logical(int value);
+  void set_by_cpa(int value);
+  void set_logical_by_cpl(int value);
   std::string to_string();
 };
